requirements.txt
asyncio==3.4.3
Flask==1.1.2
psycopg2==2.8.6
python-dotenv==0.15.0
Telethon==1.18.2